Add `create_user_group` attribute to Useradd provider
Add a `create_user_group` attribute to Chef::Provider::User::Useradd.
The default value is `nil`, so there's no change in existing behavior.
But setting it to `true` or `false` will append the `-U` or `-N`
optional arguments, respectively, to the `useradd` commandline.

This may or may not be compatible with all useradd(1) variants (it
works on Ubuntu 12+ and CentOS 6+, at least), so additional OS/distro
compatibility testing may be desirable.