rewrite linu_user provider check_lock
this removes the kinda brittle kinda insane parsing of the installed
rpm version.  now we're just less strict about the exit 1 code and
fail on the output not being what is expected.

this also likely fixes a bug in rhel in why-run mode where shell_out!
would have failed on the user-not-found and raised an exception that
would have ended the run.  the code before certainly looks buggy.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>