use threads to workaround Process.spawn buginess
there's a WONTFIX closed bug in ruby-lang.org on this bug:

https://bugs.ruby-lang.org/issues/10583

they seem to be favoring speed over accuracy with the result
that if you're doing I/O synchronization (like our run_lock?)
inside of a Process.spawn chain like this that you have to
wrap the spawn calls inside of threads.

probably a best practice to wrap every spawn call with a
Thread.

this probably also explains why we had the silly-long 20
second timeout?

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>