ignore unknown metadata fields in metadata.rb
if we're just using chef-zero there's no metadata.json file so
chef-client doesn't do json parsing (which ignores unknown metadata
fields already) but parses metadata.json which explodes.

i believe since TK goes through berks it renders the rb to json before
copying to the virt so we don't see this in TK or in running chef-client
from a real server.

this only affects the chef-client -z or chef-zolo mode of operation (and
probably legacy chef-solo)

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>