Add back compatibility with old property names
In yum 3.0 we changed the properties in the documentation, but never threw deprecation notices. Since there was no notice you would have no way to know to use the new properties. We need to add back compatibility with the old names and if we really want to kill them throw a proper deprecation warning.

Signed-off-by: Tim Smith <tsmith@chef.io>