Convert attributes defined as literal arrays
VividMash, unlike Mash, does not recursively convert values inside
arrays, just the array itself.
As such, hashes inside arrays are not converted to *Mash, introducing
a behaviour that differs from Chef 11, e.g.

having this attr defined

node['foo'] = [ { 'attr' => 'value' } ]

in Chef 11 `node[:foo][0][:attr]` returns `'value'` while in Chef 12 it
returns `nil`

This commit fixes VividMash so arrays are recursively converted and the
behaviour restored.

Plus, immutablize merge_all result, so to be sure that hashes contained
in Arrays are properly immutablized as well, so not existing attributes
are not automatically vivified (thus returning an empty VividMash
instead of nil)

This problem was first reported in chef/chef#2871.

Signed-off-by: Giacomo Bagnoli <gbagnoli@gmail.com>