Raise error if suse group manager adds non-existent user
The expected behavior of chef, as defined by our tests, is that if a
user specified in a group resource does not exist than chef should raise
an error. On suse, we use groupmod which will not raise an error if you
attempt to add a non-existent user to a group. This change adds in a
simple check that will raise an error if the specified does not exist.

Signed-off-by: Tom Duffield <tom@chef.io>