Attributes: add tests for regression in 12.0 fixed recently
Tests for the issue that led up to

https://github.com/chef/chef/issues/5447

The Hash tests actually pass against 12.13.37 so it appearly that
only broke the Immutable array objects in 12.0.  This both explains
why we didn't have any existing tests that caught it, and why it went
uncaught for so long.  It also adds some ammunition to the argument
that fixing this in 12.14.x was not a major breaking change to users
that had been relying on the buggy 12.0 behavior.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>