Enable data collector w/o token for solo, but require explicit URL
Third-party data collector implementations are supported and may not
require a token for authentication. Therefore we must allow the case
where the token is nil in Chef Solo mode. To reduce the chance that the
data collector is accidentally enabled, only set the data collector URL
to a default value in Chef Client mode.

Signed-off-by: Daniel DeLeo <dan@chef.io>