Update changelog generator to better fill our pattern
Configure the changelog generator to better fit our current pattern of
changelogs. By utilizing the HISTORY.md file, we can generate a combined
changelog of the latest version from HISTORY.md to now. Then, when we
cut a release we can simply move our current changelog into
HISTORY.md (using the appropriate rake task) and start the process over
again.

This might even be a first step to having the Chef Versioner generate
the changelog automatically for us each time it bumps the version, and
then we just need to groom it as we go along.

Signed-off-by: Tom Duffield <tom@chef.io>