Remove the Dev Tools component
This PR removes the Dev Tools component. All responsibilities of Dev
Tools would thus be handled by the Core Maintainers.  As part of this
change, I propose that

  Steven Danna, and
  Joshua Timberman

be added as Core Maintainers. All other members of Dev Tools are already
core maintainers. This move simplifies the MAINTAINER list and makes it
easier to find reviewers for the components previously handled by Dev
Tools.

To add the new core maintainers this PR requires approval from a
majority of @chef/client-core and @thom.

No procedure is specified to remove a component. I recommend that we
require a super-majority of 80% (4/5ths) of existing Dev
Tools (@chef/client-dev-tools) maintainers to approve with no vetos.

Signed-off-by: Steven Danna <steve@chef.io>