Data Collector debug log should output JSON
After changing Data Collector to use Chef::ServerAPI or
Chef::HTTP::SimpleJSON, the message output in the debug log is a ruby
inspect of the message hash rather than the JSON output that will be
sent to the data collector. Having this as JSON makes it easier to debug
and reproduce Data Collector-related issues.

Signed-off-by: Adam Leff <adam@leff.co>