Use tduffield branch of github-changelog-generator
This branch fixes a number of bugs in the changelog generator, and
supports the type of behavior that we want when generating our
changelog. There is a PR open against
upstream (skywinder/github-changelog-generator#453) to merge this
upstream.

As part of this, we can now generate the changelog on every commit.

Signed-off-by: Tom Duffield <tom@chef.io>