Update ohai resource to new style, stop overwriting name property
If #5606 is approved, any resource that creates a property that is an
already-existing Ruby method on a Chef::Resource object will throw a
deprecation warning. The ohai resource currently creates a `name`
property which conflicts with the `name` property created in the base
Chef::Resource class.

This change updates the ohai resource and provider to the new style of
defining resources and stops overwriting the `name` method.

Signed-off-by: Adam Leff <adam@leff.co>