Style/NumericPredicate comparison style
Apparently rubocop thinks the default behavior should be that we have to
rewrite every use of `foo == 0` into `foo.zero?` which is a big pile of
NOPE for me.  After discovering that `.zero?` is actually slower, I'd
prefer to go the other direction.  Same for `positive?` and
`negative?`.

These are the only uses of `zero?` in the chef/chef codebase, while I'm
pretty sure the inverse rule would touch nearly every file.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>