Chef-13 break:  script and all its subclasses now errors on 'command' property
This was always a coding bug.  The `command` property is only used
internally but is exposed for users to twiddle because script
subclasses execute.  Thus this still violates Liskov Substitution and
this is not an is-a relationship where inheritance might be appropriate,
because the script resource should /use/ the execute resource and
not /be an/ execute resource.

The more you know...

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>