Verify data bag exists before trying to create it
[ZD-12613]
There are users who want to customize permissions to prevent a group of
users from creating new data bags but want to allow the group of users
to create data bag items within an existing data bag.

If the user doesn't have permission to create data bags then the current
API request sequence will prevent the user from creating a data bag item
even if the data bag already exists.

This commit fixes the above problem by verifying (not trying to create)
if the data bag exist through a GET request, if it doesn't the it tries
to generate it, then it proceeds to create the data bag item.

Signed-off-by: Salim Afiune <afiune@chef.io>