remove most of supports API
still there on service (where it makes some sense)

also still on mount (because i have no idea if that is actively being
used or if it makes any sense at all).

converts it to a property on mount + service as well.

also removed setting it as an array -- did we ever document that and/or
does anyone use it?  i'm not religiously against that way of setting
it, but if nobody ever used it i'd rather remove the YAGNI.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>