Chef-13: Simplify DSL creation
we did actually have tests around creating actions with spaces and
hyphens in their names.  we had fallback code for properties, but it
was broken and threw a "private method define_resource called" and then
it created a closure and didn't have the `name` in the binding so it
called itself and blew up the stack.

this change walks that back.  it does prove that we still support
UTF-8 in property names, actions and DSL names, which i'd argue is
more than enough.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>