Chef-13: Support nameless resources and remove deprecated multi-arg resources
This makes nameless resources work in the DSL:

```ruby
apt_update
```

It does it by giving it an empty-string name of ""

It also drops support for multi-arg resources, that has been deprecated
for some time:

```ruby
some_resource "foo", "bar", "baz"
```

Note that multipackage package resources do not use multiple args, but a
single argument which is an arry:

```ruby
package [ "lsof", "strace", "tcpdump" ]
```

So this change does not affect that usage.  Multi-arg has been
deprecated for some time now and its not clear that it was ever used by
anyone.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>