Make ActionClass a class
since any ActionClass literally is-a Chef::Provider it should just
use inheritance and then we don't have to use the hacky hooks for
class methods in modules and it becomes more clear from the
ActionClass source what it is and what it does.

Bonus starts the deprecation of properties being injected into the
provider namespace.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>