fix and simplify rake bundle:* commands
applies the same kinds of fixes that i applied to rake dependencies:*
and mostly reverts to mainline usage of bundler.  abandons all the
trampolining, and uses the platforms that are already locked in
the Gemfile.lock and unwinds the multi-platform stuff that bundler
does natively now.

removes the BundleUtil module entirely since it is no longer used.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>