Chef-13:  restore log_location in client.rb
This was probably a bit too aggressive.  There are people who pass
logger objects in client.rb directly into the log_location, and
we'd break them, and there's no general solution.  We have a couple
of the popular configs that can be passed in as symbols, but not
the general case.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>