Update package resource with default for timeout.
Currently the timeout attribute is inconsistent in behavior between windows and non-windows platforms. This was initially observed in chef-ingredient with allowing the resource to be of type NilClass. Converges would fail on Windows but succeed on other platforms. This PR moves the default value to the package superclass and removes the attribute from windows_package.

Signed-off-by: Jennifer Davis <iennae@gmail.com>