Always run reconfigure between runs, and when HUP'd
We can't reconfigure in a trap context because we might open files or another
activity that could cause a deadlock. Ruby 2.0+ now prevents this by raising a
ThreadError.

This now runs reconfigure after every daemonized run, but still takes a HUP in
case you need to reconfigure while sleeping.

Fixes #4578

Signed-off-by: Bryan McLellan <btm@loftninjas.org>