Discard caches for files not in the manifest
cookbook_segment returns nil if the cache is not found in the manifest.
This can happen when chef client 12 runs on a node where chef client 13
generated cookbook cache files.

Currently, chef client fails in this case. This patch fixes it.

It should be safe to remove those cache files because they are not
cached in Chef 12, after all.

Signed-off-by: Akira Kitada <akitada@gmail.com>