simplify omnibus config and greenify builds again
this is also necessary for bundler-1.14.x

i'm still not entirely clear why we ever needed all the fussy software gem
configs or what the build-chef / build-chef-gem infrastructure ever
did for us.  it seems to have been mostly micro-optimization around
building the software gems before bundle installing the project in order
to take advantage of git caching.  i aggressively don't care about that,
this is quite fast enough.  we can install nokogiri and libgecode early
and that should take care of 98% of the build optimization issue.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>