Prefer Systemd with sysvinit script over Upstart for service provider
On newer Debian based systems that use systemd to have a package install a SysV init and Upstart script.  Systemd can use systemd service files or SysV init scripts.  However, on the Chef serivce provider side, this is not reflected.  So in this case, Upstart will be prefered even if Upstart itself is not installed on the system.

Signed-off-by: Grant Ridder <shortdudey123@gmail.com>