fix retries on JSON POST requests when negotiating protocol version
on the first pass through the JSON middleware it encodes the
body.  if there's a retry, it re-encodes the body as a string
with all its metacharacters escaped.

this is a particular issue when doing a first request that is
a POST that requires negotiating the API version.  when doing a
GET it isn't a problem because there's no body payload -- but a
POST or a PUT which requires a retry will get garbled and will
cause a 500.

this happens on hosted right now if trying to POST with a v2
API since hosted is only v1, so there's a retry to downgrade.

i also made the same kind of changes to the streaming download
requests, but since they're GETs its unclear to me if there was
any impact there -- but middleware could have been double-mangling
headers on a retry.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>