Set explicit page size for every search request
The previous code would advance the search start parameter by either
the user-provided rows parameter or (as a fallback) the number of rows
returned in that page.  Most code that calls the search() function
does not set rows, so the fallback is used often.

Since the search index may have stale results, it is possible to get
responses with a completely empty page. For example:

    {
       "start": 0,
       "total": 1,
       "rows": []
    }

In this case, if no rows parameter was passed, the search function
would infinitely recurse (eventually running out of stack).

To avoid this, we set rows to 1000 if no user-provided value was set.
This is the default page size of Chef Server so for most users it will
not be changing the default behavior.

Signed-off-by: Steven Danna <steve@chef.io>