Don't catch SIGCHLD from dnf_helper.py
dnf_helper.py was configured to catch the SIGCHLD signal and quit if it
was received.  In Fedora26 I have found that the fill_sack call in the
dnf python package may shell out causing a SIHCHLD call.  This results
in no information as the helper just exits before finding a package.

This removes the helper restoring the default handling of SIGCHLD which
is to ingore it.

Obvious fix.

Fixes: 6393

Signed-off-by: Brandon Bennett <bbennett@fb.com>