dnf_resource: be more specific for rhel packages
With #6351 and #6377, the DNF package provider was fixed to use yum by
default on RHEL 7, but the resource was still calling the DNF provider
if it happened to be installed.

We want to use the default of yum for RHEL 7, and only use DNF if we
call `set_priority_map` directly.

Signed-off-by: Naomi Reeves <naomi.c.reeves@gmail.com>

Resolves: Unexpectedly switching package managers