only rhel >= 8 and fedora >= 22 get dnf
the shell_out approach to parse the version string from rpm is a bit of
a failed experiment.

the shell_out that gets incurred on every package provider is a bit
terrible for performance.

DNF < 1.00 has also never formally landed in any distribution and its
very difficult at this point to deploy it.

when amazon deploys DNF we should add a version comparison for that.

If this patch causes issues we can go back to adding some form of

`provides :package ... { which("dnf" }`

That will be much faster than having the shell_out().

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>