Ensure package (un)locking is idempotent
Due to the nature of the comparison done, the `package_locked` methods
will always return false as they presume the package name is a string,
when it is in fact always coerced into an array.

Additionally in situations where the package_name is set rather than
being inherited from the package resource's name, it will always return
false.

Resolves #6361
Resolves #6493

Signed-off-by: Andy Wagner <andy@andywagner.ca>