Avoid the need to parse nil versions in the package provider
If we don't have a version installed yet we shouldn't go do a version
comparison with nil. No matter what that's a waste of time and in
dpkg/apt it's going to require shelling out now. Instead just do the
install.

This also adds the same version comparison method to the dpkg resource.
At the moment it doesn't look like this is going to get called since our
upgrade action there is a bit broken (doesn't actually compare
versions), but that's going to get fixed next.

This also cleans up the version comparison to make sure we're
always dealing with strings

Signed-off-by: Tim Smith <tsmith@chef.io>