speedup lwrp tests
reset the global resource and provider handler maps before each and
every test.

note that mutation through the top-level keys will get preserved so
this is not perfect, but deep-duping through every single test we run
becomes expensive.

this is sufficient to flush out the global state of all the test lwrp fixtures
that we setup.

it discovered several usage bugs where we either relied on this feature
or where we were testing different fixtures from the ones we thought we
were.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>