Fix knife status to show seconds when needed #5055
When a node.save operation is executed and a "knife status" is
immediately executed I get "xx minutes" instead of "xx seconds".

The code behind 'knife status' isn't designed to have the ability to
format the time difference in seconds.

To remediate issue, make sure there is a condition in place to format
the time difference in seconds when appropriate.

Signed-off-by: Chibuikem Amaechi <cramaechi@me.com>