Escape single quotes in validation_client_name #5130
When you have a single quote in the validation_client_name of knife.rb
it doesn't get escaped properly when running knife bootstrap.

In the reporter's case, he had literally set the validation_client_name
to "A path that doesn't exist" because of another bug he was running
into with some code.

Signed-off-by: Chibuikem Amaechi <cramaechi@me.com>