fix double-logging bug
somehow we're getting an IO object here which is a STDOUT wired up to
file descriptor 10 which is not the same as "STDOUT" which is wired up
to file descriptor 1.

i can't track down where or how this is happening.

this works around the problem by just inspecting if the log_location is
set to an IO object which is a tty or not which should be broadly
equivalent (and perhaps more correct than the old code?  not sure if
the edge cases matter or not).

also i'm not smart enough right now to figure out how to test this
adequately, particularly in ways that would actually catch if it
really breaks in the future (particularly due to the gnarliness of
wanting to have a functional test which constructs a tty when run on
travis without a tty....)

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>