Actually bump the version to 14 and dep on chef-config / ohai 14
Without this the kitchen tests fail as there isn't a chef-config build at ~14.0. Ohai has been bumped to 14 and the chef-config dep updated to 14. For now we have to use builds from master since we don't have a gem. I left notes that we can remove those in the future.

Signed-off-by: Tim Smith <tsmith@chef.io>