Stop mixlib-cli default clobbering mixlib-config settings
The pre-14 precedence level is:

1. mixlib-cli setting
2. mixlib-cli default
3. mixlib-config setting
4. mixlib-config default

This means that if an option has a mixlib-cli default that it cannot
ever be set in the config file.

This PR swaps 2+3 around:

1. mixlib-cli setting
2. mixlib-config setting
3. mixlib-cli default
4. mixlib-config default

Now the mixlib-cli defaults still take precedence over mixlib-config
defaults, but it is possible to set a value in config.rb if there's
a mixlib-cli default setting (which creeps into the settings in hidden
ways if you just use `boolean: true` in mixlib-cli).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>